code has been refactorated

+++ src/PLDMS.Core/Entities/TaskLanguage.cs
@@ -4,7 +4,7 @@ namespace PLDMS.Core.Entities;
 
 public class TaskLanguage
 {
-    public int TaskId { get; set; }
+    public long TaskId { get; set; }
     public Task Task { get; set; }
     
     public ProgrammingLanguage ProgrammingLanguage { get; set; }
